Add files via upload
Client asked me to write a routine that verifies a database contains valid data.  Our state's drivers license numbers is based on an algorithm explained in the file soundex1.jpg.  test-data.csv contains four records, two of which are encoded correctly, two of them are not.  The .ipynb file is the Jupyter Notebook for this project which shows the checkpoints used to construct the code.  The dl-verify-b.py script checks the records and reports which are valid using a Boolean test.